+++ index.html
@@ -232,7 +232,7 @@
     </template>
 
     <!-- Подключение файла JavaScript -->
-    <script src="./js/main.js"></script>
+    <script src="./js/main.js" type="module"></script>
 
 </body>
